fix: Add publicPath option to HtmlWebpack plugin
When running the build, HtmlWebpack plugin added a "/" to start
of the src attribute in script.

E.g. <script src="/bundle.js"></script>

As a result, an error was shown in the console stating that it
couldn't find "https://anordinaryusername.github.io/bundle.js".

Include publicPath with "" to remove the "/" and fix the error.