feat: Add production options to config
Change module.exports to a function for env variable to check if
the build is for production. Use it to check for filename's
hash mode and to disable source-maps.

When building for prod, inline-source-maps take up a lot of space.
Disable it for production only.

Use hash-mode for caching. Look here for an explanation
(https://stackoverflow.com/a/59197260).