refactor: Move classes into separate files
Move the classes inside InstructionSet.ts into their own files for
readability.

Change the generated instruction percentages in generateProblem().
Previously, it generated negative percentages which is bad. Conditionally
change the bounds of percentages so that it never becomes negative.

Update InstructionSet.test.ts to account for classC presence. The
test threw an error since classC is undefined and only appears 50%
of the time.

Update relative path in InstructionSetForm.tsx to reflect path changes.